requirements.txt
asgiref==3.8.1
Django==5.2.3
django-ckeditor==6.7.3
django-js-asset==3.1.2
psycopg==3.2.9
psycopg-binary==3.2.9
sqlparse==0.5.3
typing_extensions==4.14.0
